requirements.txt
beautifulsoup4==4.13.3
faiss_cpu==1.10.0
nltk==3.9.1
numpy
pandas==2.2.3
pdf2image==1.17.0
pdfplumber==0.11.5
pytesseract==0.3.13
Requests==2.32.3
scikit_learn==1.4.1.post1
scipy==1.15.2
sentence_transformers==3.4.1
tqdm==4.67.1
urllib3==2.3.0
